+++ external/fv3fit/fv3fit/train_microphysics.py
@@ -53,10 +53,9 @@ from fv3fit.emulation.transforms import (
     CloudWaterDiffPrecpd,
     GscondClassesV1,
     GscondClassesV1OneHot,
-    TendencyToFlux,
-    MoistStaticEnergyTransform,
     GscondRoute,
 )
+from fv3fit.emulation.flux import TendencyToFlux, MoistStaticEnergyTransform
 
 from fv3fit.emulation.layers.normalization import standard_deviation_all_features
 from fv3fit.wandb import (
